Create Update-original-record
This query will search for a matching record on the other table and, if found, will update it rather than create a new record. By setting the Condition line, the UI Buttons will switch between Update and Transfer depending on if there is anything in the "Originating MEDINC Number" field on the form.